api/requirements.txt
fastapi==0.110.1
mangum==0.17.0
motor==3.3.2
pymongo==4.6.0
pydantic>=2.6.4
python-dotenv>=1.0.1
email-validator>=2.2.0
starlette>=0.27.0
uvicorn>=0.25.0
aiosmtplib>=3.0.0